moved test from QueryTest to SnappySQLQuerySuite and fixed occasional failures due to incorrect hashCode()
