+++ cluster/src/test/scala/org/apache/spark/sql/SnappySQLQuerySuite.scala
@@ -38,10 +38,12 @@ package org.apache.spark.sql
 import io.snappydata.{Property, SnappyFunSuite}
 import org.scalatest.Matchers._
 
-import org.apache.spark.sql.catalyst.encoders.RowEncoder
+import org.apache.spark.sql.catalyst.encoders.{ExpressionEncoder, RowEncoder}
+import org.apache.spark.sql.execution.FilterExec
 import org.apache.spark.sql.execution.aggregate.{HashAggregateExec, SnappyHashAggregateExec}
 import org.apache.spark.sql.execution.benchmark.ColumnCacheBenchmark
-import org.apache.spark.sql.functions.rand
+import org.apache.spark.sql.execution.joins.HashJoinExec
+import org.apache.spark.sql.functions.{bround, rand, round}
 import org.apache.spark.sql.test.SQLTestData.TestData2
 
 
@@ -49,6 +51,7 @@ import org.apache.spark.sql.test.SQLTestData.TestData2
 class SnappySQLQuerySuite extends SnappyFunSuite {
 
   private lazy val session: SnappySession = snc.snappySession
+  private val idPattern = "(,[0-9]+)?#[0-9L]+".r
 
   // Ported test from Spark
   test("SNAP-1885 : left semi greater than predicate and equal operator") {
@@ -214,7 +217,7 @@ class SnappySQLQuerySuite extends SnappyFunSuite {
     assert(plan.find(_.isInstanceOf[SnappyHashAggregateExec]).isDefined)
     assert(plan.find(_.isInstanceOf[HashAggregateExec]).isEmpty)
     // collect the result to force default hashAggregateSize property take effect
-    implicit val encoder = RowEncoder(rs.schema)
+    implicit val encoder: ExpressionEncoder[Row] = RowEncoder(rs.schema)
     val result = session.createDataset(rs.collect().toSeq)
     session.sql(s"set ${Property.HashAggregateSize} = -1")
     try {
@@ -404,6 +407,164 @@ class SnappySQLQuerySuite extends SnappyFunSuite {
           "(a is not null AND b is null)").collect())
     assert(df.selectExpr("a IS DISTINCT FROM b").collect() === df.selectExpr(
       "NOT (a <=> b)").collect())
+  }
+
+  test("Push down TPCH Q19") {
+    session.sql("set spark.sql.autoBroadcastJoinThreshold=-1")
+    try {
+      for (_ <- 1 to 4) {
+        testTPCHQ19()
+      }
+    } finally {
+      session.sql(s"set spark.sql.autoBroadcastJoinThreshold=${10L * 1024 * 1024}")
+    }
+  }
+
+  private def testTPCHQ19(): Unit = {
+    // check common sub-expression elimination in query leading to push down
+    // of filters should not be inhibited due to ParamLiterals
+    import session.implicits._
+
+    session.sql("create table ct1 (id long, data string) using column")
+    session.sql("create table ct2 (id long, data string) using column")
+    session.sql("insert into ct1 select id, 'data' || id from range(100000)")
+    session.sql("insert into ct2 select id, 'data' || id from range(100000)")
+
+    var ds = session.sql("select ct1.id, ct2.data from ct1 join ct2 on (ct1.id = ct2.id) where " +
+        "(ct1.id < 1000 and ct2.data = 'data100') or (ct1.id < 1000 and ct1.data = 'data100')")
+    var analyzedFilter = "Filter (((ID#0 < cast(ParamLiteral:0#0,1000 as bigint)) && " +
+        "(DATA#0 = ParamLiteral:1#0,data100)) || ((ID#0 < cast(ParamLiteral:0#0,1000 as " +
+        "bigint)) && (DATA#0 = ParamLiteral:1#0,data100)))"
+
+    def expectedTree: String =
+      s"""Project [ID#0, DATA#0]
+         |+- $analyzedFilter
+         |   +- Join Inner, (ID#0 = ID#0)
+         |      :- SubqueryAlias CT1
+         |      :  +- Relation[ID#0,DATA#0] ColumnFormatRelation[APP.CT1]
+         |      +- SubqueryAlias CT2
+         |         +- Relation[ID#0,DATA#0] ColumnFormatRelation[APP.CT2]
+         |""".stripMargin
+
+    assert(idPattern.replaceAllIn(ds.queryExecution.analyzed.treeString, "#0") === expectedTree)
+    assert(ds.collect() === Array(Row(100L, "data100")))
+
+    // check filter push down in the plan
+    var filters = ds.queryExecution.executedPlan.collect {
+      case f: FilterExec => assert(f.child.nodeName === "ColumnTableScan"); f
+    }
+    assert(filters.length === 2)
+    assert(filters.forall(_.toString.contains("<")))
+    // check pushed down filters should not be in HashJoin
+    var joins = ds.queryExecution.executedPlan.collect {
+      case j: HashJoinExec => j
+    }
+    assert(joins.length === 1)
+    assert(joins.head.condition.isDefined)
+    var condString = joins.head.condition.get.toString()
+    assert(condString.contains("DATA#"))
+    assert(!condString.contains("ID#"))
+
+    // similar query but different values in the two positions should lead to a different
+    // plan with no filter push down
+    ds = session.sql("select ct1.id, ct2.data from ct1 join ct2 on (ct1.id = ct2.id) where " +
+        "(ct1.id < 1000 and ct2.data = 'data100') or (ct1.id < 20 and ct1.data = 'data100')")
+    analyzedFilter = "Filter (((ID#0 < cast(ParamLiteral:0#0,1000 as bigint)) && " +
+        "(DATA#0 = ParamLiteral:1#0,data100)) || ((ID#0 < cast(ParamLiteral:2#0,20 as " +
+        "bigint)) && (DATA#0 = ParamLiteral:1#0,data100)))"
+    assert(idPattern.replaceAllIn(ds.queryExecution.analyzed.treeString, "#0") === expectedTree)
+    assert(ds.collect() === Array(Row(100L, "data100")))
+
+    // check no filter push down in the plan
+    filters = ds.queryExecution.executedPlan.collect {
+      case f: FilterExec => assert(f.child.nodeName === "ColumnTableScan"); f
+    }
+    assert(filters.length === 2)
+    assert(filters.forall(!_.toString.contains("<")))
+    // check all filters should be in HashJoin
+    joins = ds.queryExecution.executedPlan.collect {
+      case j: HashJoinExec => j
+    }
+    assert(joins.length === 1)
+    assert(joins.head.condition.isDefined)
+    condString = joins.head.condition.get.toString()
+    assert(condString.contains("DATA#"))
+    assert(condString.contains("ID#"))
+
+    ds = session.sql("select ct1.id, ct2.data from ct1 join ct2 on (ct1.id = ct2.id) where " +
+        "(ct1.id < 10 and ct2.data = 'data100') or (ct1.id < 10 and ct1.data = 'data100')")
+    assert(ds.collect().length === 0)
+    // check filter push down in the plan
+    filters = ds.queryExecution.executedPlan.collect {
+      case f: FilterExec => assert(f.child.nodeName === "ColumnTableScan"); f
+    }
+    assert(filters.length === 2)
+    assert(filters.forall(_.toString.contains("<")))
+    ds = session.sql("select ct1.id, ct2.data from ct1 join ct2 on (ct1.id = ct2.id) where " +
+        "(ct1.id < 10 and ct2.data = 'data100') or (ct1.id < 20 and ct1.data = 'data100')")
+    assert(ds.collect().length === 0)
+    // check no filter push down in the plan
+    filters = ds.queryExecution.executedPlan.collect {
+      case f: FilterExec => assert(f.child.nodeName === "ColumnTableScan"); f
+    }
+    assert(filters.length === 2)
+    assert(filters.forall(!_.toString.contains("<")))
+
+    session.sql("drop table ct1")
+    session.sql("drop table ct2")
+
+    // check for some combinations of repeated constants
+    val df = Seq(5, 55, 555).map(Tuple1(_)).toDF("a")
+    checkAnswer(
+      df.select(round('a), round('a, -1), round('a, -2)),
+      Seq(Row(5, 10, 0), Row(55, 60, 100), Row(555, 560, 600))
+    )
+    checkAnswer(
+      df.select(bround('a), bround('a, -1), bround('a, -2)),
+      Seq(Row(5, 0, 0), Row(55, 60, 100), Row(555, 560, 600))
+    )
+
+    val pi = BigDecimal("3.1415")
+    checkAnswer(
+      session.sql(s"SELECT round($pi, -3), round($pi, -2), round($pi, -1), " +
+          s"round($pi, 0), round($pi, 1), round($pi, 2), round($pi, 3)"),
+      Seq(Row(BigDecimal("0E3"), BigDecimal("0E2"), BigDecimal("0E1"), BigDecimal(3),
+        BigDecimal("3.1"), BigDecimal("3.14"), BigDecimal("3.142")))
+    )
+    checkAnswer(
+      session.sql(s"SELECT bround($pi, -3), bround($pi, -2), bround($pi, -1), " +
+          s"bround($pi, 0), bround($pi, 1), bround($pi, 2), bround($pi, 3)"),
+      Seq(Row(BigDecimal("0E3"), BigDecimal("0E2"), BigDecimal("0E1"), BigDecimal(3),
+        BigDecimal("3.1"), BigDecimal("3.14"), BigDecimal("3.142")))
+    )
+
+    // more than 4 constants to replace
+    val pi1 = pi + 1
+    val pi2 = pi + 2
+    val pi3 = pi + 3
+    val pi4 = pi + 4
+    val pi5 = pi + 5
+    val pi6 = pi + 6
+    checkAnswer(
+      session.sql(s"SELECT round($pi, -3), round($pi6, -2), round($pi, -1), " +
+          s"round($pi, 0), round($pi1, 1), round($pi, 2), round($pi2, 3), " +
+          s"round($pi3, 1), round($pi1, 2), round($pi4, 2), round($pi, 3), " +
+          s"round($pi5, 3), round($pi4, 0), round($pi, 1), round($pi6, 2)"),
+      Seq(Row(BigDecimal("0E3"), BigDecimal("0E2"), BigDecimal("0E1"),
+        BigDecimal(3), BigDecimal("4.1"), BigDecimal("3.14"), BigDecimal("5.142"),
+        BigDecimal("6.1"), BigDecimal("4.14"), BigDecimal("7.14"), BigDecimal("3.142"),
+        BigDecimal("8.142"), BigDecimal(7), BigDecimal("3.1"), BigDecimal("9.14")))
+    )
+    checkAnswer(
+      session.sql(s"SELECT bround($pi, -3), bround($pi6, -2), bround($pi, -1), " +
+          s"bround($pi, 0), bround($pi1, 1), bround($pi, 2), bround($pi2, 3), " +
+          s"bround($pi3, 1), bround($pi1, 2), bround($pi4, 2), bround($pi, 3), " +
+          s"bround($pi5, 3), bround($pi4, 0), bround($pi, 1), bround($pi6, 2)"),
+      Seq(Row(BigDecimal("0E3"), BigDecimal("0E2"), BigDecimal("0E1"),
+        BigDecimal(3), BigDecimal("4.1"), BigDecimal("3.14"), BigDecimal("5.142"),
+        BigDecimal("6.1"), BigDecimal("4.14"), BigDecimal("7.14"), BigDecimal("3.142"),
+        BigDecimal("8.142"), BigDecimal(7), BigDecimal("3.1"), BigDecimal("9.14")))
+    )
   }
 }
